requirements.txt
tensorflow>=1.15
keras>=2.2.4
numpy
scikit-learn
biopython
imblearn
matplotlib